+++ example/lib/features_modal/modal_confirm.dart
@@ -86,10 +86,14 @@ class _FeaturesModalConfirmState extends State<FeaturesModalConfirm> {
                   TextButton(
                     child: Text('OK (${state.selection?.length ?? 0})'),
                     style: ButtonStyle(
-                      backgroundColor: MaterialStateProperty.all<Color>(Theme.of(context).primaryColor),
-                      textStyle: MaterialStateProperty.all<TextStyle>(TextStyle(color: Colors.white)),
+                      backgroundColor: MaterialStateProperty.all<Color>(
+                          Theme.of(context).primaryColor),
+                      textStyle: MaterialStateProperty.all<TextStyle>(
+                          TextStyle(color: Colors.white)),
                     ),
-                    onPressed: (state.selection?.isValid ?? true) ? () => state.closeModal(confirmed: true) : null,
+                    onPressed: (state.selection?.isValid ?? true)
+                        ? () => state.closeModal(confirmed: true)
+                        : null,
                   ),
                 ],
               ),
@@ -132,8 +136,10 @@ class _FeaturesModalConfirmState extends State<FeaturesModalConfirm> {
                 child: TextButton(
                   child: const Text('Send'),
                   style: ButtonStyle(
-                    backgroundColor: MaterialStateProperty.all<Color>(Colors.redAccent),
-                    textStyle: MaterialStateProperty.all<TextStyle>(TextStyle(color: Colors.white)),
+                    backgroundColor:
+                        MaterialStateProperty.all<Color>(Colors.redAccent),
+                    textStyle: MaterialStateProperty.all<TextStyle>(
+                        TextStyle(color: Colors.white)),
                   ),
                   onPressed: () => state.closeModal(confirmed: true),
                 ),
