+++ lib/src/choices_empty.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 /// Default widget for empty choices list
 class S2ChoicesEmpty extends StatelessWidget {
   /// Default constructor
-  const S2ChoicesEmpty({Key? key}): super(key: key);
+  const S2ChoicesEmpty({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
