+++ lib/src/tile/tile.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import '../widget.dart';
+import 'package:flutter_awesome_select/flutter_awesome_select.dart';
 
 /// Default trigger/tile widget
 class S2Tile<T> extends StatelessWidget {
@@ -129,8 +129,8 @@ class S2Tile<T> extends StatelessWidget {
   })  : title = title ?? state.titleWidget,
         value = value ?? Text(state.selected.toString()),
         onTap = onTap ?? state.showModal,
-        isLoading = isLoading ?? state.selected?.isResolving ?? false,
-        isError = isError ?? state.selected?.isNotValid ?? false,
+        isLoading = isLoading ?? state.selected.isResolving,
+        isError = isError ?? state.selected.isNotValid,
         super(key: key);
 
   /// Returns default trailing widget
