+++ pubspec.yaml
@@ -1,7 +1,7 @@
 name: flutter_awesome_select
 description: Easily convert your usual single/multiple select or dropdown into page, dialog, or bottom sheet with various choice input such as radio, checkbox, switch, chips, or custom widget.
 homepage: https://github.com/artflutter/flutter_awesome_select
-version: 6.4.0
+version: 6.5.0-beta
 
 environment:
   sdk: '>=2.17.0 <3.0.0'
